<refactor> modify: 移除AStarUtil逻辑至AStarEntity
