--- Assets/com.gamearki.pathfinding/PureRuntime/AStarUtil.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e90549212f395ad42acb65d2113512a4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
